requirements.txt
PyJWT>=2.0.0,<3.0.0
